simplificacao Triangle
