+++ chimera/render/partition/Triangle.hpp
@@ -1,6 +1,5 @@
 #pragma once
-#include "chimera/render/VertexData.hpp"
-#include <list>
+#include <glm/glm.hpp>
 
 namespace Chimera {
 
@@ -26,8 +25,4 @@ struct Triangle {
     bool splitter;
     glm::vec3 normal;
 };
-
-void triangleFromVertexDataIndex(VertexData* vertexData, uint32_t* indexData, const uint32_t& indexSize, std::list<Triangle*>& vTris);
-void triangleFromVertexData(VertexData* vertexData, const uint32_t& vertexSize, std::list<Triangle*>& vTris);
-// bool tringleListIsConvex(std::vector<VertexData>& vertexList, std::vector<Triangle*>& _vTriangle);
 } // namespace Chimera
